assessment_crud_done

+++ patient/models.py
@@ -42,7 +42,7 @@ class Assessment(BaseModel):
     assessment_type = models.CharField(max_length=100, null=True, blank=True)
     assessment_date = models.DateField(null=True, blank=True)
     questions_answers = models.CharField(max_length=100, null=True, blank=True)
-    final_score = models.FloatField(null=True, blank=True)
+    final_score = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
 
     class Meta:
         db_table = 'assessment'
